Stop unnecessarily using uint64 variable for uint32 data part 2
UserHistoryPredictor::LearningSegmentFingerprint returns uint32 data.  We don't need to use set<uint64> there.
No behavior change is intended.

BUG=none
TEST=unittest

git-svn-id: https://mozc.googlecode.com/svn/trunk@335 a6090854-d499-a067-5803-1114d4e51264

+++ src/prediction/user_history_predictor.cc
@@ -1737,7 +1737,7 @@ void UserHistoryPredictor::InsertHistory(bool is_suggestion_selected,
   MakeLearningSegments(*segments, &learning_segments);
 
   string all_key, all_value;
-  set<uint64> seen;
+  set<uint32> seen;
   bool this_was_seen = false;
   const size_t history_segments_size =
       learning_segments.history_segments_size();
